Reorganise auth and transaction directories
