+++ src/actions/auth/index.js
@@ -1,7 +1,6 @@
 import pick from 'lodash/pick';
-import authApi from '../api/authClient';
-import dataApi from '../api/dataClient';
-import history from '../history';
+import authApi from '../../api/authClient';
+import history from '../../history';
 import {
   REGISTRATION_REQUEST_SENT,
   REGISTRATION_SUCCESSFUL,
@@ -14,15 +13,7 @@ import {
   ACCOUNT_UPDATE_REQUEST_SENT,
   ACCOUNT_UPDATE_SUCCESSFUL,
   AUTHENTICATION_ERROR,
-  FETCH_TRANSACTIONS,
-  SHOW_TRANSACTION_MODAL,
-  HIDE_TRANSACTION_MODAL,
-  CREATE_TRANSACTION_REQUEST_SENT,
-  CREATE_TRANSACTION_SUCCESSFUL,
-  DELETE_TRANSACTION_REQUEST_SENT,
-  DELETE_TRANSACTION_SUCCESSFUL,
-  TRANSACTION_ERROR
-} from './types';
+} from './types'
 
 const ACCEPTED_JWT_HEADERS = [
   'access-token',
@@ -84,9 +75,7 @@ export const logoutUser = () => async (dispatch) => {
   try {
     await authApi.delete('/auth/sign_out', { headers: auth_headers })
 
-    dispatch({
-      type: SIGN_OUT_SUCCESSFUL
-    });
+    dispatch({ type: SIGN_OUT_SUCCESSFUL });
 
     localStorage.removeItem('auth_headers')
     history.push('/login')
@@ -146,75 +135,6 @@ export const updateUser = formValues => async dispatch => {
       payload: error.response.data.errors
     });
   }
-}
-
-export const fetchTransactions = () => async dispatch => {
-  try {
-    const response = await dataApi.get('/transactions');
-
-    dispatch({
-      type: FETCH_TRANSACTIONS,
-      payload: response.data.data
-    })
-
-  } catch(error) {
-    dispatch({
-      type: TRANSACTION_ERROR,
-      payload: error.response
-    });
-  }
-}
-
-export const showTransactionModal = () => {
-  return(
-    { type: SHOW_TRANSACTION_MODAL }
-  )
-}
-
-export const hideTransactionModal = () => {
-  return(
-    { type: HIDE_TRANSACTION_MODAL }
-  )
-}
-
-export const createTransaction = formValues => async dispatch => {
-  dispatch({ type: CREATE_TRANSACTION_REQUEST_SENT })
-
-  try {
-    const response = await dataApi.post('/transactions', formValues);
-
-    dispatch({
-      type: CREATE_TRANSACTION_SUCCESSFUL,
-      payload: response.data.data
-    });
-
-    dispatch({ type: HIDE_TRANSACTION_MODAL });
-
-  } catch(error) {
-    dispatch({
-      type: TRANSACTION_ERROR,
-      payload: error.response
-    });
-  }
-}
-
-export const deleteTransaction = transactionId => async dispatch => {
-  dispatch({ type: DELETE_TRANSACTION_REQUEST_SENT })
-
-    try {
-      const response = await dataApi.delete(`/transactions/${transactionId}`);
-
-      dispatch({
-        type: DELETE_TRANSACTION_SUCCESSFUL,
-        payload: transactionId
-      });
-
-    } catch(error) {
-      dispatch({
-        type: TRANSACTION_ERROR,
-        payload: error.response
-      });
-    }
 }
 
 const storeAuthHeaders = (responseHeaders) => {
