+++ src/actions/auth/types.js
@@ -14,16 +14,3 @@ export const ACCOUNT_UPDATE_REQUEST_SENT = 'ACCOUNT_UPDATE_REQUEST_SENT';
 export const ACCOUNT_UPDATE_SUCCESSFUL = 'ACCOUNT_UPDATE_SUCCESSFUL';
 
 export const AUTHENTICATION_ERROR = 'AUTHENTICATION_ERROR';
-
-export const FETCH_TRANSACTIONS = 'FETCH_TRANSACTIONS';
-
-export const SHOW_TRANSACTION_MODAL = 'SHOW_TRANSACTION_MODAL';
-export const HIDE_TRANSACTION_MODAL = 'HIDE_TRANSACTION_MODAL';
-
-export const CREATE_TRANSACTION_REQUEST_SENT = 'CREATE_TRANSACTION_REQUEST_SENT';
-export const CREATE_TRANSACTION_SUCCESSFUL = 'CREATE_TRANSACTION_SUCCESSFUL';
-
-export const DELETE_TRANSACTION_REQUEST_SENT = 'DELETE_TRANSACTION_REQUEST_SENT';
-export const DELETE_TRANSACTION_SUCCESSFUL = 'DELETE_TRANSACTION_SUCCESSFUL';
-
-export const TRANSACTION_ERROR = 'TRANSACTION_ERROR';
