+++ src/components/Portfolio.js
@@ -3,7 +3,7 @@ import { Button, Divider } from 'semantic-ui-react';
 import { connect } from 'react-redux';
 import TransactionList from './transactions/TransactionList';
 import TransactionCreate from './transactions/TransactionCreate';
-import { showTransactionModal } from '../actions/';
+import { showTransactionModal } from '../actions/transactions';
 
 const Portfolio = (props) => {
   return (
