+++ src/components/auth/LoginForm.js
@@ -2,18 +2,19 @@ import React from 'react';
 import { Button, Form, Grid, Header, Message, Segment } from 'semantic-ui-react';
 import { Field, reduxForm } from 'redux-form';
 import { connect } from 'react-redux';
-import { registerUser } from '../actions';
+import { Link } from 'react-router-dom';
+import { loginUser } from '../../actions/auth';
 
-class SignUpForm extends React.Component {
+class LoginForm extends React.Component {
   onSubmit = (formValues) => {
-    this.props.registerUser(formValues)
+    this.props.loginUser(formValues)
   }
 
   render () {
     return(
       <Grid textAlign='center' style={{ paddingTop: '5%' }} verticalAlign='middle'>
         <Grid.Column style={{ maxWidth: 450 }}>
-          <Header as='h2' color='teal' textAlign='center'> Sign up for an account </Header>
+          <Header as='h2' color='teal' textAlign='center'> Login to account </Header>
           <Form error size='large' onSubmit={this.props.handleSubmit(this.onSubmit)}>
             <Segment stacked>
               <Field
@@ -26,25 +27,8 @@ class SignUpForm extends React.Component {
               />
               <Field
                 component={Form.Input}
-                name='name'
-                placeholder='Full Name'
-                fluid icon='user circle'
-                iconPosition='left'
-                type='text'
-              />
-              <Field
-                component={Form.Input}
                 name='password'
                 placeholder='Password'
-                fluid
-                icon='lock'
-                iconPosition='left'
-                type='password'
-              />
-              <Field
-                component={Form.Input}
-                name='password_confirmation'
-                placeholder='Confirm Password'
                 fluid
                 icon='lock'
                 iconPosition='left'
@@ -55,9 +39,15 @@ class SignUpForm extends React.Component {
                 color='teal'
                 fluid size='large'
                 loading={this.props.isLoading}
-              > Sign Up
+              > Login
               </Button>
 
+              <Message>
+                New to us?
+                <Link to='/signup'>
+                  Sign Up
+                </Link>
+              </Message>
               <Message
                 color='orange'
                 error
@@ -79,8 +69,8 @@ const mapStateToProps = (state) => {
 }
 
 export default reduxForm({
-  form: 'registration',
+  form: 'login',
 })(connect(
   mapStateToProps,
-  { registerUser }
-)(SignUpForm));
+  { loginUser }
+)(LoginForm));
