+++ src/components/auth/SignUpForm.js
@@ -2,19 +2,18 @@ import React from 'react';
 import { Button, Form, Grid, Header, Message, Segment } from 'semantic-ui-react';
 import { Field, reduxForm } from 'redux-form';
 import { connect } from 'react-redux';
-import { Link } from 'react-router-dom';
-import { loginUser } from '../actions';
+import { registerUser } from '../../actions/auth';
 
-class LoginForm extends React.Component {
+class SignUpForm extends React.Component {
   onSubmit = (formValues) => {
-    this.props.loginUser(formValues)
+    this.props.registerUser(formValues)
   }
 
   render () {
     return(
       <Grid textAlign='center' style={{ paddingTop: '5%' }} verticalAlign='middle'>
         <Grid.Column style={{ maxWidth: 450 }}>
-          <Header as='h2' color='teal' textAlign='center'> Login to account </Header>
+          <Header as='h2' color='teal' textAlign='center'> Sign up for an account </Header>
           <Form error size='large' onSubmit={this.props.handleSubmit(this.onSubmit)}>
             <Segment stacked>
               <Field
@@ -27,8 +26,25 @@ class LoginForm extends React.Component {
               />
               <Field
                 component={Form.Input}
+                name='name'
+                placeholder='Full Name'
+                fluid icon='user circle'
+                iconPosition='left'
+                type='text'
+              />
+              <Field
+                component={Form.Input}
                 name='password'
                 placeholder='Password'
+                fluid
+                icon='lock'
+                iconPosition='left'
+                type='password'
+              />
+              <Field
+                component={Form.Input}
+                name='password_confirmation'
+                placeholder='Confirm Password'
                 fluid
                 icon='lock'
                 iconPosition='left'
@@ -39,15 +55,9 @@ class LoginForm extends React.Component {
                 color='teal'
                 fluid size='large'
                 loading={this.props.isLoading}
-              > Login
+              > Sign Up
               </Button>
 
-              <Message>
-                New to us?
-                <Link to='/signup'>
-                  Sign Up
-                </Link>
-              </Message>
               <Message
                 color='orange'
                 error
@@ -69,8 +79,8 @@ const mapStateToProps = (state) => {
 }
 
 export default reduxForm({
-  form: 'login',
+  form: 'registration',
 })(connect(
   mapStateToProps,
-  { loginUser }
-)(LoginForm));
+  { registerUser }
+)(SignUpForm));
