+++ src/components/dashboard/MarketStatistics.js
