--- src/components/transactions/TransactionDelete.js
@@ -1,11 +1,0 @@
-import React from 'react';
-
-const TransactionDelete = () => {
-  return (
-    <div>
-      TransactionDelete
-    </div>
-  )
-}
-
-export default TransactionDelete;
