--- src/components/transactions/TransactionEdit.js
@@ -1,11 +1,0 @@
-import React from 'react';
-
-const TransactionEdit = () => {
-  return (
-    <div>
-      TransactionEdit
-    </div>
-  )
-}
-
-export default TransactionEdit;
